Tester for steg1
